+++ subiquity/controllers/keyboard.py
@@ -15,9 +15,9 @@
 
 import logging
 
+from subiquitycore.async_helpers import schedule_task
 from subiquitycore.controller import BaseController
 
-from subiquity.async_helpers import schedule_task
 from subiquity.models.keyboard import KeyboardSetting
 from subiquity.ui.views import KeyboardView
