+++ subiquity/controllers/refresh.py
@@ -19,12 +19,12 @@ import os
 
 import requests.exceptions
 
+from subiquitycore.async_helpers import (
+    schedule_task,
+    )
 from subiquitycore.controller import BaseController
 from subiquitycore.core import Skip
 
-from subiquity.async_helpers import (
-    schedule_task,
-    )
 
 log = logging.getLogger('subiquity.controllers.refresh')
